Every function in its own file, part 2 [#CLICKHOUSE-2]
