+++ dbms/src/Functions/finalizeAggregation.cpp
@@ -1,0 +1,71 @@
+#include <Functions/IFunction.h>
+#include <Functions/FunctionFactory.h>
+#include <Functions/FunctionHelpers.h>
+#include <DataTypes/DataTypeAggregateFunction.h>
+#include <Columns/ColumnAggregateFunction.h>
+#include <Common/typeid_cast.h>
+
+
+namespace DB
+{
+
+namespace ErrorCodes
+{
+    extern const int ILLEGAL_COLUMN;
+    extern const int ILLEGAL_TYPE_OF_ARGUMENT;
+}
+
+
+/** finalizeAggregation(agg_state) - get the result from the aggregation state.
+  * Takes state of aggregate function. Returns result of aggregation (finalized state).
+  */
+class FunctionFinalizeAggregation : public IFunction
+{
+public:
+    static constexpr auto name = "finalizeAggregation";
+    static FunctionPtr create(const Context &)
+    {
+        return std::make_shared<FunctionFinalizeAggregation>();
+    }
+
+    String getName() const override
+    {
+        return name;
+    }
+
+    size_t getNumberOfArguments() const override
+    {
+        return 1;
+    }
+
+    DataTypePtr getReturnTypeImpl(const DataTypes & arguments) const override
+    {
+        const DataTypeAggregateFunction * type = checkAndGetDataType<DataTypeAggregateFunction>(arguments[0].get());
+        if (!type)
+            throw Exception("Argument for function " + getName() + " must have type AggregateFunction - state of aggregate function.",
+                ErrorCodes::ILLEGAL_TYPE_OF_ARGUMENT);
+
+        return type->getReturnType();
+    }
+
+    void executeImpl(Block & block, const ColumnNumbers & arguments, size_t result, size_t /*input_rows_count*/) override
+    {
+        const ColumnAggregateFunction * column_with_states
+            = typeid_cast<const ColumnAggregateFunction *>(&*block.getByPosition(arguments.at(0)).column);
+        if (!column_with_states)
+            throw Exception("Illegal column " + block.getByPosition(arguments.at(0)).column->getName()
+                    + " of first argument of function "
+                    + getName(),
+                ErrorCodes::ILLEGAL_COLUMN);
+
+        block.getByPosition(result).column = column_with_states->convertToValues();
+    }
+};
+
+
+void registerFunctionFinalizeAggregation(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionFinalizeAggregation>();
+}
+
+}
