+++ dbms/src/Functions/runningDifference.cpp
@@ -1,0 +1,13 @@
+#include <Functions/runningDifference.h>
+#include <Functions/FunctionFactory.h>
+
+
+namespace DB
+{
+
+void registerFunctionRunningDifference(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionRunningDifferenceImpl<true>>();
+}
+
+}
