+++ dbms/src/Functions/runningDifferenceStartingWithFirstValue.cpp
@@ -1,0 +1,13 @@
+#include <Functions/runningDifference.h>
+#include <Functions/FunctionFactory.h>
+
+
+namespace DB
+{
+
+void registerFunctionRunningDifferenceStartingWithFirstValue(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionRunningDifferenceImpl<false>>();
+}
+
+}
