+++ dbms/src/Functions/version.cpp
@@ -1,0 +1,48 @@
+#include <Functions/IFunction.h>
+#include <Functions/FunctionFactory.h>
+#include <DataTypes/DataTypeString.h>
+#include <Common/config_version.h>
+
+
+namespace DB
+{
+
+/** version() - returns the current version as a string.
+  */
+class FunctionVersion : public IFunction
+{
+public:
+    static constexpr auto name = "version";
+    static FunctionPtr create(const Context &)
+    {
+        return std::make_shared<FunctionVersion>();
+    }
+
+    String getName() const override
+    {
+        return name;
+    }
+
+    size_t getNumberOfArguments() const override
+    {
+        return 0;
+    }
+
+    DataTypePtr getReturnTypeImpl(const DataTypes & /*arguments*/) const override
+    {
+        return std::make_shared<DataTypeString>();
+    }
+
+    void executeImpl(Block & block, const ColumnNumbers &, size_t result, size_t input_rows_count) override
+    {
+        block.getByPosition(result).column = DataTypeString().createColumnConst(input_rows_count, String(VERSION_STRING));
+    }
+};
+
+
+void registerFunctionVersion(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionVersion>();
+}
+
+}
